requirements.txt
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
rich==13.7.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
requests==2.31.0